当たり判定警告解除

+++ collision/Collision.cpp
@@ -25,13 +25,13 @@ bool Collision::CircleCollision(const float& X1, const float& Y1, const float& R
 
 
 bool Collision::SphereCollision(const float& X1, const float& Y1, const float& Z1, const float& R1, const float& X2, const float& Y2, const float& Z2, const float& R2) {
-	int a = X1 - X2;
-	int b = Y1 - Y2;
-	int c = Z1 - Z2;
+	float a = X1 - X2;
+	float b = Y1 - Y2;
+	float c = Z1 - Z2;
 	//2‚Â‚Ì‹——£‚ğŒvZ
-	int distance = sqrtf(a * a + b * b + c * c);
+	float distance = sqrtf(a * a + b * b + c * c);
 	//”¼Œa‚Ì‡Œv‚ğŒvZ
-	int radius = R1 + R2;
+	float radius = R1 + R2;
 	if (distance <= radius) {
 		return true;
 	} else {
